+++ addon/components/sortable-group.js
@@ -1,12 +1,18 @@
-import Ember from 'ember';
+import { A } from '@ember/array';
+import Component from '@ember/component';
+import { set, get } from '@ember/object';
+import { run } from '@ember/runloop';
 import layout from '../templates/components/sortable-group';
 import computed from 'ember-new-computed';
-const { A, Component, get, set, run } = Ember;
+import { invokeAction } from 'ember-invoke-action';
+
 const a = A;
 const NO_MODEL = {};
 
 export default Component.extend({
   layout: layout,
+
+  attributeBindings: ['data-test-selector'],
 
   /**
     @property direction
@@ -149,9 +155,9 @@ export default Component.extend({
     });
 
     if (groupModel !== NO_MODEL) {
-      this.sendAction('onChange', groupModel, itemModels, draggedModel);
+      invokeAction(this, 'onChange', groupModel, itemModels, draggedModel);
     } else {
-      this.sendAction('onChange', itemModels, draggedModel);
+      invokeAction(this, 'onChange', itemModels, draggedModel);
     }
   }
 });