+++ addon/helpers/drag.js
@@ -1,5 +1,5 @@
-import Ember from 'ember';
-const { $ } = Ember;
+import { registerAsyncHelper } from '@ember/test';
+import $ from 'jquery';
 
 /**
   Drags elements by an offset specified in pixels.
@@ -52,8 +52,13 @@ export function drag(app, mode, itemSelector, offsetFn, callbacks = {}) {
     let item = findWithAssert(itemSelector);
     let itemOffset = item.offset();
     let offset = offsetFn();
-    let targetX = itemOffset.left + offset.dx;
-    let targetY = itemOffset.top + offset.dy;
+    let itemElement = item.get(0);
+    let rect = itemElement.getBoundingClientRect();
+    let scale = itemElement.clientHeight / (rect.bottom - rect.top);
+    let halfwayX = itemOffset.left + (offset.dx * scale) / 2;
+    let halfwayY = itemOffset.top + (offset.dy * scale) / 2;
+    let targetX = itemOffset.left + offset.dx * scale;
+    let targetY = itemOffset.top + offset.dy * scale;
 
     triggerEvent(app, item, start, {
       pageX: itemOffset.left,
@@ -73,6 +78,11 @@ export function drag(app, mode, itemSelector, offsetFn, callbacks = {}) {
     if (callbacks.dragmove) {
       andThen(callbacks.dragmove);
     }
+
+    triggerEvent(app, item, move, {
+      pageX: halfwayX,
+      pageY: halfwayY
+    });
 
     triggerEvent(app, item, move, {
       pageX: targetX,
@@ -99,4 +109,4 @@ function triggerEvent(app, el, type, props) {
   });
 }
 
-export default Ember.Test.registerAsyncHelper('drag', drag);
+export default registerAsyncHelper('drag', drag);